+++ lib/src/crypto/encryptor.dart
@@ -1,10 +1,5 @@
 import 'package:encrypt/encrypt.dart';
 
-enum EncryptionAlgorithm {
-  aes,
-  salsa20,
-}
-
 abstract class Encryptor {
   Encrypted encrypt(String key, String plainText);