+++ lib/src/enc_shared_pref.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 import 'package:encrypt_shared_preferences/src/crypto/aes.dart';
-import 'package:encrypt_shared_preferences/src/crypto/salsa20.dart';
 import 'package:encrypt_shared_preferences/src/shared_preferences_decorator.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
@@ -12,7 +11,6 @@ class EncryptedSharedPreferences {
   static String? _key;
 
   static late SharedPreferencesDecorator _decorator;
-  static late Encryptor _encryptor;
 
   static final EncryptedSharedPreferences _instance =
       EncryptedSharedPreferences._();
@@ -24,19 +22,11 @@ class EncryptedSharedPreferences {
 
   Stream<String> get stream => _decorator.listenable.stream;
 
-  static Future<void> initialize(String key,
-      {EncryptionAlgorithm? algorithm = EncryptionAlgorithm.aes}) async {
+  static Future<void> initialize(String key) async {
     _key = key;
-    if (algorithm == EncryptionAlgorithm.salsa20) {
-      _encryptor = Salsa20Encryptor();
-    }
-    if (algorithm == EncryptionAlgorithm.aes) {
-      _encryptor = AESEncryptor();
-    }
-
     _decorator = SharedPreferencesDecorator(
         preferences: await SharedPreferences.getInstance(),
-        encryptor: _encryptor,
+        encryptor: AESEncryptor(),
         key: _key!);
   }